requirements.txt
filelock==3.13.4
fsspec==2024.3.1
iniconfig==2.0.0
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
nada-dsl==0.1.1
networkx==3.3
numpy==1.26.4
packaging==24.0
pluggy==1.4.0
py-nillion-client==0.1.1
pytest==8.1.1
pytest-asyncio==0.23.6
python-dotenv==1.0.0
scikit-learn==1.4.2
scipy==1.13.0
sympy==1.12
threadpoolctl==3.4.0
torch==2.2.2
typing_extensions==4.11.0
